Rename to peers behaviour
